rename: 연산자 상수, type 위치 Caluator로 수정
